@@ -1,20 +1,24 @@
 apply plugin: 'com.android.library'
-apply plugin: 'com.github.dcendents.android-maven'
 apply plugin: 'com.jfrog.bintray'
 
-version = "1.1.7.18"
+version = "1.1.7.22"
 
 android {
-    compileSdkVersion 26
-    buildToolsVersion '27.0.3'
+    namespace "za.co.adyo.android"
+
+    compileSdkVersion 34
 
     defaultConfig {
-        minSdkVersion 19
-        targetSdkVersion 26
-        versionCode 2
-        versionName "1.1.7.18"
+        minSdkVersion 21
+        targetSdkVersion 34
+        versionCode 3
+        versionName "1.1.7.22"
 
+        // Required for our library version retrieval to work as of Android Studio 4.1.0: https://issuetracker.google.com/issues/158695880
+        buildConfigField 'int', 'VERSION_CODE', "${project.android_version_code}"
+        buildConfigField 'String', 'VERSION_NAME', "\"${project.android_version_name}\""
     }
+
     buildTypes {
         release {
             minifyEnabled false
