@@ -9,10 +9,9 @@ import android.content.Intent;
 import android.graphics.PorterDuff;
 import android.graphics.drawable.Drawable;
 import android.net.Uri;
-import android.support.annotation.Nullable;
-import android.support.annotation.StyleRes;
-import android.support.customtabs.CustomTabsIntent;
-import android.support.v4.content.ContextCompat;
+import androidx.annotation.Nullable;
+import androidx.annotation.StyleRes;
+import androidx.browser.customtabs.CustomTabsIntent;
 import android.util.Log;
 import android.webkit.WebSettings;
 import android.webkit.WebView;
