@@ -1,6 +1,6 @@
 package za.co.adyo.android.models;
 
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 
 import org.json.JSONArray;
 import org.json.JSONObject;
