@@ -8,8 +8,8 @@ import android.graphics.PorterDuff;
 import android.graphics.drawable.Drawable;
 import android.os.Build;
 import android.os.Bundle;
-import android.support.annotation.RequiresApi;
-import android.support.annotation.StyleRes;
+import androidx.annotation.RequiresApi;
+import androidx.annotation.StyleRes;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.webkit.WebResourceRequest;
