@@ -1,6 +1,6 @@
-package za.co.adyo.adyo.android;
+package za.co.adyo.android;
 
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
 
 import za.co.adyo.android.requests.PlacementRequestParams;
